requirements.txt
dnspython==1.16.0
falcon==1.4.1
gunicorn==19.9.0
pymongo==3.7.2
python-dateutil==2.7.5
python-mimeparse==1.6.0
pytz==2018.7
schema==0.6.8
six==1.12.0
